+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2021-04-07 
+#### (compatibility: 2020.1 - 2021.1.*) 
 ### Added
 - Can now select a default Ahk sdk from the Ahk settings
 - Syntax highlighting now supports hotkeys